+++ clang/unittests/StaticAnalyzer/CallDescriptionTest.cpp
@@ -100,30 +100,30 @@ public:
 TEST(CallEvent, CallDescription) {
   // Test simple name matching.
   EXPECT_TRUE(tooling::runToolOnCode(
-      new CallDescriptionAction({
+      std::unique_ptr<CallDescriptionAction>(new CallDescriptionAction({
           {{"bar"}, false}, // false: there's no call to 'bar' in this code.
           {{"foo"}, true},  // true: there's a call to 'foo' in this code.
-      }), "void foo(); void bar() { foo(); }"));
+      })), "void foo(); void bar() { foo(); }"));
 
   // Test arguments check.
   EXPECT_TRUE(tooling::runToolOnCode(
-      new CallDescriptionAction({
+      std::unique_ptr<CallDescriptionAction>(new CallDescriptionAction({
           {{"foo", 1}, true},
           {{"foo", 2}, false},
-      }), "void foo(int); void foo(int, int); void bar() { foo(1); }"));
+      })), "void foo(int); void foo(int, int); void bar() { foo(1); }"));
 
   // Test lack of arguments check.
   EXPECT_TRUE(tooling::runToolOnCode(
-      new CallDescriptionAction({
+      std::unique_ptr<CallDescriptionAction>(new CallDescriptionAction({
           {{"foo", None}, true},
           {{"foo", 2}, false},
-      }), "void foo(int); void foo(int, int); void bar() { foo(1); }"));
+      })), "void foo(int); void foo(int, int); void bar() { foo(1); }"));
 
   // Test qualified names.
   EXPECT_TRUE(tooling::runToolOnCode(
-      new CallDescriptionAction({
+      std::unique_ptr<CallDescriptionAction>(new CallDescriptionAction({
           {{{"std", "basic_string", "c_str"}}, true},
-      }),
+      })),
       "namespace std { inline namespace __1 {"
       "  template<typename T> class basic_string {"
       "  public:"
@@ -138,18 +138,18 @@ TEST(CallEvent, CallDescription) {
 
   // A negative test for qualified names.
   EXPECT_TRUE(tooling::runToolOnCode(
-      new CallDescriptionAction({
+      std::unique_ptr<CallDescriptionAction>(new CallDescriptionAction({
           {{{"foo", "bar"}}, false},
           {{{"bar", "foo"}}, false},
           {{"foo"}, true},
-      }), "void foo(); struct bar { void foo(); }; void test() { foo(); }"));
+      })), "void foo(); struct bar { void foo(); }; void test() { foo(); }"));
 
   // Test CDF_MaybeBuiltin - a flag that allows matching weird builtins.
   EXPECT_TRUE(tooling::runToolOnCode(
-      new CallDescriptionAction({
+      std::unique_ptr<CallDescriptionAction>(new CallDescriptionAction({
           {{"memset", 3}, false},
           {{CDF_MaybeBuiltin, "memset", 3}, true}
-      }),
+      })),
       "void foo() {"
       "  int x;"
       "  __builtin___memset_chk(&x, 0, sizeof(x),"